+++ capa/features/extractors/ida/file.py
@@ -27,8 +27,20 @@ import capa.features.extractors.helpers
 import capa.features.extractors.strings
 import capa.features.extractors.ida.helpers
 from capa.features.file import Export, Import, Section, FunctionName
-from capa.features.common import FORMAT_PE, FORMAT_ELF, Format, String, Feature, Characteristic
-from capa.features.address import NO_ADDRESS, Address, FileOffsetAddress, AbsoluteVirtualAddress
+from capa.features.common import (
+    FORMAT_PE,
+    FORMAT_ELF,
+    Format,
+    String,
+    Feature,
+    Characteristic,
+)
+from capa.features.address import (
+    NO_ADDRESS,
+    Address,
+    FileOffsetAddress,
+    AbsoluteVirtualAddress,
+)
 
 MAX_OFFSET_PE_AFTER_MZ = 0x200
 
@@ -52,7 +64,9 @@ def check_segment_for_pe(seg: idaapi.segment_t) -> Iterator[tuple[int, int]]:
     todo = []
     for mzx, pex, i in mz_xor:
         # find all segment offsets containing XOR'd "MZ" bytes
-        for off in capa.features.extractors.ida.helpers.find_byte_sequence(seg.start_ea, seg.end_ea, mzx):
+        for off in capa.features.extractors.ida.helpers.find_byte_sequence(
+            seg.start_ea, seg.end_ea, mzx
+        ):
             todo.append((off, mzx, pex, i))
 
     while len(todo):
@@ -64,7 +78,10 @@ def check_segment_for_pe(seg: idaapi.segment_t) -> Iterator[tuple[int, int]]:
         if seg_max < (e_lfanew + 4):
             continue
 
-        newoff = struct.unpack("<I", capa.features.extractors.helpers.xor_static(idc.get_bytes(e_lfanew, 4), i))[0]
+        newoff = struct.unpack(
+            "<I",
+            capa.features.extractors.helpers.xor_static(idc.get_bytes(e_lfanew, 4), i),
+        )[0]
 
         # assume XOR'd "PE" bytes exist within threshold
         if newoff > MAX_OFFSET_PE_AFTER_MZ:
@@ -85,7 +102,9 @@ def extract_file_embedded_pe() -> Iterator[tuple[Feature, Address]]:
         - '-R' from console
         - Check 'Load resource sections' when opening binary in IDA manually
     """
-    for seg in capa.features.extractors.ida.helpers.get_segments(skip_header_segments=True):
+    for seg in capa.features.extractors.ida.helpers.get_segments(
+        skip_header_segments=True
+    ):
         for ea, _ in check_segment_for_pe(seg):
             yield Characteristic("embedded pe"), FileOffsetAddress(ea)
 
@@ -97,7 +116,11 @@ def extract_file_export_names() -> Iterator[tuple[Feature, Address]]:
         if forwarded_name is None:
             yield Export(name), AbsoluteVirtualAddress(ea)
         else:
-            forwarded_name = capa.features.extractors.helpers.reformat_forwarded_export_name(forwarded_name)
+            forwarded_name = (
+                capa.features.extractors.helpers.reformat_forwarded_export_name(
+                    forwarded_name
+                )
+            )
             yield Export(forwarded_name), AbsoluteVirtualAddress(ea)
             yield Characteristic("forwarded export"), AbsoluteVirtualAddress(ea)
 
@@ -118,7 +141,9 @@ def extract_file_import_names() -> Iterator[tuple[Feature, Address]]:
         if info[1] and info[2]:
             # e.g. in mimikatz: ('cabinet', 'FCIAddFile', 11L)
             # extract by name here and by ordinal below
-            for name in capa.features.extractors.helpers.generate_symbols(info[0], info[1], include_dll=True):
+            for name in capa.features.extractors.helpers.generate_symbols(
+                info[0], info[1], include_dll=True
+            ):
                 yield Import(name), addr
             dll = info[0]
             symbol = f"#{info[2]}"
@@ -131,7 +156,9 @@ def extract_file_import_names() -> Iterator[tuple[Feature, Address]]:
         else:
             continue
 
-        for name in capa.features.extractors.helpers.generate_symbols(dll, symbol, include_dll=True):
+        for name in capa.features.extractors.helpers.generate_symbols(
+            dll, symbol, include_dll=True
+        ):
             yield Import(name), addr
 
     for ea, info in capa.features.extractors.ida.helpers.get_file_externs().items():
@@ -145,7 +172,9 @@ def extract_file_section_names() -> Iterator[tuple[Feature, Address]]:
         - '-R' from console
         - Check 'Load resource sections' when opening binary in IDA manually
     """
-    for seg in capa.features.extractors.ida.helpers.get_segments(skip_header_segments=True):
+    for seg in capa.features.extractors.ida.helpers.get_segments(
+        skip_header_segments=True
+    ):
         yield Section(idaapi.get_segm_name(seg)), AbsoluteVirtualAddress(seg.start_ea)
 
 
@@ -161,10 +190,10 @@ def extract_file_strings() -> Iterator[tuple[Feature, Address]]:
 
         # differing to common string extractor factor in segment offset here
         for s in capa.features.extractors.strings.extract_ascii_strings(seg_buff):
-            yield String(s.s), FileOffsetAddress(seg.start_ea + s.offset)
+            yield String(s.s), AbsoluteVirtualAddress(seg.start_ea + s.offset)
 
         for s in capa.features.extractors.strings.extract_unicode_strings(seg_buff):
-            yield String(s.s), FileOffsetAddress(seg.start_ea + s.offset)
+            yield String(s.s), AbsoluteVirtualAddress(seg.start_ea + s.offset)
 
 
 def extract_file_function_names() -> Iterator[tuple[Feature, Address]]: